+++ Neo4j.Driver/Neo4j.Driver/Internal/IO/MessageSerializers/V5/FailureMessageSerializer.cs
@@ -1,0 +1,33 @@
+﻿// Copyright (c) 2002-2022 "Neo4j,"
+// Neo4j Sweden AB [http://neo4j.com]
+// 
+// This file is part of Neo4j.
+// 
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+// 
+//     http://www.apache.org/licenses/LICENSE-2.0
+// 
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+using Neo4j.Driver.Internal.Messaging;
+
+namespace Neo4j.Driver.Internal.IO.MessageSerializers.V5
+{
+    internal class FailureMessageSerializer : V3.FailureMessageSerializer
+    {
+        public override object Deserialize(IPackStreamReader reader, byte signature, long size)
+        {
+            var values = reader.ReadMap();
+            var code = values["code"]?.ToString();
+            var message = values["message"]?.ToString();
+
+            return new FailureMessage(code, message);
+        }
+    }
+}
