+++ Neo4j.Driver/Neo4j.Driver/Internal/Protocol/BoltProtocolV3MessageFormat.cs
@@ -66,7 +66,7 @@ namespace Neo4j.Driver.Internal.Protocol
             AddHandler<ResetMessageSerializer>();
 
             // Response Message Types
-            AddHandler<FailureMessageV1Serializer>();
+            AddHandler<FailureMessageSerializer>();
             AddHandler<IgnoredMessageSerializer>();
             AddHandler<RecordMessageSerializer>();
             AddHandler<SuccessMessageSerializer>();
