+++ Neo4j.Driver/Neo4j.Driver/Internal/Protocol/BoltProtocolV5_0MessageFormat.cs
@@ -15,8 +15,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-using Neo4j.Driver.Internal.IO.MessageSerializers;
 using Neo4j.Driver.Internal.IO.ValueSerializers;
+using FailureMessageSerializer = Neo4j.Driver.Internal.IO.MessageSerializers.V5.FailureMessageSerializer;
 
 namespace Neo4j.Driver.Internal.Protocol
 {
@@ -33,7 +33,7 @@ namespace Neo4j.Driver.Internal.Protocol
             RemoveHandler<UnboundRelationshipSerializer>();
             AddHandler<ElementUnboundRelationshipSerializer>();
 
-            RemoveHandler<FailureMessageV1Serializer>();
+            RemoveHandler<IO.MessageSerializers.V3.FailureMessageSerializer>();
             AddHandler<FailureMessageSerializer>();
         }
     }
